+++ packages/core/src/store/__tests__/state-store-integration.test.ts
@@ -2,7 +2,7 @@ import { jest, expect, it, beforeEach, describe, test } from '@jest/globals'
 import tmp from 'tmp-promise'
 import {
   AnchorStatus,
-  CommitType,
+  EventType,
   StreamState,
   IpfsApi,
   SignatureStatus,
@@ -51,7 +51,7 @@ describe('Level data store', () => {
     metadata: { controllers: ['foo'] },
     signature: SignatureStatus.GENESIS,
     anchorStatus: AnchorStatus.NOT_REQUESTED,
-    log: [{ cid: FAKE_CID, type: CommitType.GENESIS }],
+    log: [{ cid: FAKE_CID, type: EventType.INIT }],
   }
 
   beforeEach(async () => {
@@ -82,7 +82,7 @@ describe('Level data store', () => {
   it('pins not anchored stream correctly with IPFS pinning', async () => {
     const state: StreamState = {
       ...streamState,
-      log: [{ cid: FAKE_CID, type: CommitType.GENESIS }],
+      log: [{ cid: FAKE_CID, type: EventType.INIT }],
     }
     await expect(store.stateStore.load(streamId)).resolves.toBeNull()
     const pinSpy = jest.spyOn(store.pinning, 'pin')