+++ packages/stream-tests/src/__tests__/caip10/tezos.test.ts
@@ -1,18 +1,17 @@
 import { jest } from '@jest/globals'
 import { createIPFS } from '@ceramicnetwork/ipfs-daemon'
-import { IpfsApi } from '@ceramicnetwork/common'
+import { CeramicApi, IpfsApi } from '@ceramicnetwork/common'
 import { clearDid, happyPath, wrongProof } from './caip-flows.js'
 import { TezosAuthProvider, TezosProvider } from '@ceramicnetwork/blockchain-utils-linking'
 import { InMemorySigner } from '@taquito/signer'
 import HttpRequestMock from 'http-request-mock'
-import type { Ceramic } from '@ceramicnetwork/core'
 
 const privateKey = 'p2sk2obfVMEuPUnadAConLWk7Tf4Dt3n4svSgJwrgpamRqJXvaYcg1'
 
 let provider: TezosProvider
 let publicKey: string
 
-let ceramic: Ceramic
+let ceramic: CeramicApi
 let ipfs: IpfsApi
 
 const mocker = HttpRequestMock.setupForUnitTest('fetch')