+++ packages/stream-tests/src/__tests__/model-instance-document.test.ts
@@ -1,6 +1,6 @@
-import { jest, test, expect, beforeAll, afterAll, describe } from '@jest/globals'
+import { jest } from '@jest/globals'
 import getPort from 'get-port'
-import { AnchorStatus, CommitType, IpfsApi } from '@ceramicnetwork/common'
+import { AnchorStatus, EventType, IpfsApi } from '@ceramicnetwork/common'
 import { Utils as CoreUtils } from '@ceramicnetwork/core'
 import { createIPFS, swarmConnect } from '@ceramicnetwork/ipfs-daemon'
 import {
@@ -161,7 +161,7 @@ describe('ModelInstanceDocument API http-client tests', () => {
     expect(doc.metadata.model.equals(midMetadata.model)).toBe(true)
     expect(doc.metadata.unique).toBeInstanceOf(Uint8Array)
     expect(doc.state.log.length).toEqual(1)
-    expect(doc.state.log[0].type).toEqual(CommitType.GENESIS)
+    expect(doc.state.log[0].type).toEqual(EventType.INIT)
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.PENDING)
     expect(doc.metadata.model.toString()).toEqual(model.id.toString())
     await expect(TestUtils.isPinned(ceramic.admin, doc.id)).resolves.toBeTruthy()
@@ -179,7 +179,7 @@ describe('ModelInstanceDocument API http-client tests', () => {
     expect(docWithRelation.metadata.model.equals(midRelationMetadata.model)).toBe(true)
     expect(docWithRelation.metadata.unique).toBeInstanceOf(Uint8Array)
     expect(docWithRelation.state.log.length).toEqual(1)
-    expect(docWithRelation.state.log[0].type).toEqual(CommitType.GENESIS)
+    expect(docWithRelation.state.log[0].type).toEqual(EventType.INIT)
     expect(docWithRelation.state.anchorStatus).toEqual(AnchorStatus.PENDING)
     expect(docWithRelation.metadata.model.toString()).toEqual(modelWithRelation.id.toString())
     await expect(TestUtils.isPinned(ceramic.admin, docWithRelation.id)).resolves.toBeTruthy()
@@ -196,8 +196,8 @@ describe('ModelInstanceDocument API http-client tests', () => {
 
     expect(doc.content).toEqual(CONTENT1)
     expect(doc.state.log.length).toEqual(2)
-    expect(doc.state.log[0].type).toEqual(CommitType.GENESIS)
-    expect(doc.state.log[1].type).toEqual(CommitType.SIGNED)
+    expect(doc.state.log[0].type).toEqual(EventType.INIT)
+    expect(doc.state.log[1].type).toEqual(EventType.DATA)
   })
 
   test(`Cannot create document with relation that isn't a valid streamid`, async () => {
@@ -257,8 +257,8 @@ describe('ModelInstanceDocument API http-client tests', () => {
 
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.ANCHORED)
     expect(doc.state.log.length).toEqual(2)
-    expect(doc.state.log[0].type).toEqual(CommitType.GENESIS)
-    expect(doc.state.log[1].type).toEqual(CommitType.ANCHOR)
+    expect(doc.state.log[0].type).toEqual(EventType.INIT)
+    expect(doc.state.log[1].type).toEqual(EventType.TIME)
     expect(doc.content).toEqual(CONTENT0)
   })
 
@@ -273,9 +273,9 @@ describe('ModelInstanceDocument API http-client tests', () => {
 
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.ANCHORED)
     expect(doc.state.log.length).toEqual(3)
-    expect(doc.state.log[0].type).toEqual(CommitType.GENESIS)
-    expect(doc.state.log[1].type).toEqual(CommitType.SIGNED)
-    expect(doc.state.log[2].type).toEqual(CommitType.ANCHOR)
+    expect(doc.state.log[0].type).toEqual(EventType.INIT)
+    expect(doc.state.log[1].type).toEqual(EventType.DATA)
+    expect(doc.state.log[2].type).toEqual(EventType.TIME)
     expect(doc.content).toEqual(CONTENT1)
   })
 
@@ -294,12 +294,12 @@ describe('ModelInstanceDocument API http-client tests', () => {
 
     expect(doc.state.anchorStatus).toEqual(AnchorStatus.ANCHORED)
     expect(doc.state.log.length).toEqual(6)
-    expect(doc.state.log[0].type).toEqual(CommitType.GENESIS)
-    expect(doc.state.log[1].type).toEqual(CommitType.SIGNED)
-    expect(doc.state.log[2].type).toEqual(CommitType.ANCHOR)
-    expect(doc.state.log[3].type).toEqual(CommitType.SIGNED)
-    expect(doc.state.log[4].type).toEqual(CommitType.SIGNED)
-    expect(doc.state.log[5].type).toEqual(CommitType.ANCHOR)
+    expect(doc.state.log[0].type).toEqual(EventType.INIT)
+    expect(doc.state.log[1].type).toEqual(EventType.DATA)
+    expect(doc.state.log[2].type).toEqual(EventType.TIME)
+    expect(doc.state.log[3].type).toEqual(EventType.DATA)
+    expect(doc.state.log[4].type).toEqual(EventType.DATA)
+    expect(doc.state.log[5].type).toEqual(EventType.TIME)
     expect(doc.content).toEqual(CONTENT3)
   })
 
@@ -321,7 +321,6 @@ describe('ModelInstanceDocument API http-client tests', () => {
     await expect(
       ModelInstanceDocument.single(ceramic, {
         ...midSingleMetadata,
-        // @ts-expect-error
         controller: { invalid: 'object' },
       })
     ).rejects.toThrow(/Attempting to create a ModelInstanceDocument with an invalid DID string/)
@@ -387,21 +386,6 @@ describe('ModelInstanceDocument API http-client tests', () => {
     const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)
     await expect(TestUtils.isPinned(ceramic.admin, doc.id)).resolves.toBeTruthy()
     await expect(TestUtils.isPinned(ceramic.admin, model.id)).resolves.toBeTruthy()
-  })
-
-  test('unindex and reindex', async () => {
-    const indexApi = core.index
-    const count = () => indexApi.count({ models: [model.id] })
-    const start = await count()
-    // Index as usual
-    const document = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)
-    await expect(count()).resolves.toEqual(start + 1)
-    // Unindex
-    await document.shouldIndex(false)
-    await expect(count()).resolves.toEqual(start)
-    // Reindex
-    await document.shouldIndex(true)
-    await expect(count()).resolves.toEqual(start + 1)
   })
 })