+++ packages/stream-tile-handler/src/__tests__/tile-document-handler.test.ts
@@ -7,7 +7,7 @@ import { TileDocument } from '@ceramicnetwork/stream-tile'
 import {
   CeramicSigner,
   CommitData,
-  CommitType,
+  EventType,
   GenesisCommit,
   IpfsApi,
   SignedCommitContainer,
@@ -233,7 +233,7 @@ describeIfV3('TileDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: commit.jws,
     }
@@ -251,7 +251,7 @@ describeIfV3('TileDocumentHandler', () => {
 
     const commitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: COMMITS.genesisGenerated.linkedBlock,
       envelope: COMMITS.genesisGenerated.jws,
     }
@@ -288,7 +288,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -309,7 +309,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply signed
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
     }
@@ -332,7 +332,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -351,7 +351,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply signed
     const signedCommitData_1 = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload1,
       envelope: signedCommit1.jws,
     }
@@ -375,7 +375,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply signed
     const signedCommitData_2 = {
       cid: FAKE_CID_3,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload2,
       envelope: signedCommit2.jws,
     }
@@ -402,7 +402,7 @@ describeIfV3('TileDocumentHandler', () => {
 
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
       timestamp: Date.now(),
@@ -425,7 +425,7 @@ describeIfV3('TileDocumentHandler', () => {
 
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -450,7 +450,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -475,7 +475,7 @@ describeIfV3('TileDocumentHandler', () => {
       // apply signed
       const signedCommitData_1 = {
         cid: FAKE_CID_2,
-        type: CommitType.SIGNED,
+        type: EventType.DATA,
         commit: sPayload,
         envelope: signedCommit.jws,
       }
@@ -498,7 +498,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -523,7 +523,7 @@ describeIfV3('TileDocumentHandler', () => {
       // apply signed
       const signedCommitData_1 = {
         cid: FAKE_CID_2,
-        type: CommitType.SIGNED,
+        type: EventType.DATA,
         commit: sPayload,
         envelope: signedCommit.jws,
       }
@@ -546,7 +546,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -571,7 +571,7 @@ describeIfV3('TileDocumentHandler', () => {
       // apply signed
       const signedCommitData_1 = {
         cid: FAKE_CID_2,
-        type: CommitType.SIGNED,
+        type: EventType.DATA,
         commit: sPayload,
         envelope: signedCommit.jws,
       }
@@ -596,7 +596,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply genesis
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
     }
@@ -617,7 +617,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply signed
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
     }
@@ -627,7 +627,7 @@ describeIfV3('TileDocumentHandler', () => {
     // apply anchor
     const anchorCommitData = {
       cid: FAKE_CID_3,
-      type: CommitType.ANCHOR,
+      type: EventType.TIME,
       commit: COMMITS.r2.commit,
       proof: COMMITS.proof,
       timestamp: 1615799679,
@@ -655,7 +655,7 @@ describeIfV3('TileDocumentHandler', () => {
     // genesis commit applied one hour before rotation
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
       timestamp: rotateDate.valueOf() / 1000 - 60 * 60,
@@ -680,7 +680,7 @@ describeIfV3('TileDocumentHandler', () => {
 
     const signedCommitData = {
       cid: FAKE_CID_2,
-      type: CommitType.SIGNED,
+      type: EventType.DATA,
       commit: sPayload,
       envelope: signedCommit.jws,
       // 24 hours after rotation
@@ -711,7 +711,7 @@ describeIfV3('TileDocumentHandler', () => {
     // commit is applied 1 hour before rotation
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
       timestamp: rotateDate.valueOf() / 1000 - 60 * 60,
@@ -743,7 +743,7 @@ describeIfV3('TileDocumentHandler', () => {
     // commit is applied 1 hour after the rotation
     const genesisCommitData = {
       cid: FAKE_CID_1,
-      type: CommitType.GENESIS,
+      type: EventType.INIT,
       commit: payload,
       envelope: genesisCommit.jws,
       timestamp: Math.floor(rotateDate.valueOf() / 1000) + 60 * 60,