+++ packages/http-client/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.1.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/http-client@4.1.0-rc.0...@ceramicnetwork/http-client@4.1.0) (2024-02-02)
+
+**Note:** Version bump only for package @ceramicnetwork/http-client
+
+
+
+
+
+# [4.1.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/http-client@4.0.0...@ceramicnetwork/http-client@4.1.0-rc.0) (2024-02-01)
+
+
+### Features
+
+* Bump major verison on all packages except common, core, and http-client ([#3139](https://github.com/ceramicnetwork/js-ceramic/issues/3139)) ([aafe0ef](https://github.com/ceramicnetwork/js-ceramic/commit/aafe0ef4187935ac7f842b3ed8c8a481e8d418bf))
+* change all commitType name and values ([#3138](https://github.com/ceramicnetwork/js-ceramic/issues/3138)) ([39d521a](https://github.com/ceramicnetwork/js-ceramic/commit/39d521a9a671964f10b8aff5585a22218c39a0d6))
+
+
+
+
+
+# [5.0.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/http-client@4.0.0...@ceramicnetwork/http-client@5.0.0) (2024-02-01)
+
+**Note:** Version bump only for package @ceramicnetwork/http-client
+
+
+
+
+
+# [4.0.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/http-client@4.0.0-rc.1...@ceramicnetwork/http-client@4.0.0) (2024-01-31)
+
+**Note:** Version bump only for package @ceramicnetwork/http-client
+
+
+
+
+
+# [4.0.0-rc.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/http-client@3.4.0-rc.0...@ceramicnetwork/http-client@4.0.0-rc.1) (2024-01-30)
+
+**Note:** Version bump only for package @ceramicnetwork/http-client
+
+
+
+
+
 # [3.4.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/http-client@3.2.0...@ceramicnetwork/http-client@3.4.0-rc.0) (2024-01-25)
 
 **Note:** Version bump only for package @ceramicnetwork/http-client